# api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: CassetteOS App Management API
  version: v2
  description: |
    CassetteOS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.
servers:
  - url: /v2/app_management
...
security:
  - access_token: []
paths:
...
  /convert:
    post:
      summary: Convert to compose app from another file format
      operationId: convert
      tags:
        - Common methods
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum:
              - appfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/ComposeAppOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
  /global:
    get:
      summary: Get all global settings
      operationId: getGlobalSettings
      tags:
        - Common methods
      responses:
        "200":
          $ref: "#/components/responses/GlobalSettingsOK"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
  /global/{key}:
    put:
      summary: Update a global setting
      operationId: updateGlobalSetting
      tags:
        - Common methods
      parameters:
        - $ref: "#/components/parameters/GlobalSettingKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalSetting"
      responses:
        "200":
          $ref: "#/components/responses/GlobalSettingOK"
        "400":
          $ref: "#/components/responses/ResponseBadRequest"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    delete:
      summary: Delete a global setting
      operationId: deleteGlobalSetting
      tags:
        - Common methods
      parameters:
        - $ref: "#/components/parameters/GlobalSettingKey"
      responses:
        "200":
          $ref: "#/components/responses/ResponseOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"

    get:
      summary: Get a global setting
      operationId: getGlobalSetting
      tags:
        - Common methods
      parameters:
        - $ref: "#/components/parameters/GlobalSettingKey"
      responses:
        "200":
          $ref: "#/components/responses/GlobalSettingOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
  /container/{id}:
    patch:
      summary: Recreate the container app
      description: |
        TODO - add description
      operationId: recreateContainerByID
      tags:
        - Container methods
      parameters:
        - $ref: "#/components/parameters/ContainerID"
        - $ref: "#/components/parameters/PullLatestImage"
        - $ref: "#/components/parameters/Force"
      responses:
        "200":
          $ref: "#/components/responses/ContainerRecreateOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    GlobalSettingKey:
      name: key
      description: Key of the global setting
      in: path
      required: true
      schema:
        type: string
        example: "OPENAI_API_KEY"
...
    ContainerID:
      name: id
      description: Container ID in the Docker runtime
      in: path
      required: true
      schema:
        type: string
      example: 09e70f4a7dd3fc43de6a1bc40d45c79f857544e13679e5b1a47409034a7ce621
...
    PullLatestImage:
      name: pull
      description: Try to pull the latest image from the registry
      in: query
      schema:
        type: boolean
        default: false
    Force:
      name: force
      in: query
      schema:
        type: boolean
        default: false
...
  responses:
    ResponseOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "OK"
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
    ResponseBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Bad Request"
...
    GlobalSettingOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/GlobalSetting"
    GlobalSettingsOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/GlobalSetting"
...
    ComposeAppOK:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - properties:
                  data:
                    $ref: "#/components/schemas/ComposeAppWithStoreInfo"
        application/yaml:
          schema:
            type: string
...
    ContainerRecreateOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
...
  schemas:
    BaseResponse:
      properties:
        message:
          description: message returned by server side if there is any
          type: string
          example: ""
...
    GlobalSetting:
      required:
        - key
        - value
      properties:
        key:
          description: key of the setting
          type: string
          readOnly: true
          example: "OPENAI_API_KEY"
        value:
          description: value of the setting
          type: string
          example: "sk-xxxxxxx"
        description:
          description: description of the setting
          type: string
          example: "OpenAI API Key"
...
    ComposeAppStoreInfo:
      required:
        - author
        - category
        - description
        - developer
        - icon
        - screenshot_link
        - tagline
        - thumbnail
        - title
        - tips
        - index
        - port_map
      properties:
        store_app_id:
          $ref: "#/components/schemas/StoreAppID"
        title:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Syncthing
        image:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: syncthing:latest
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Syncthing is a continuous file synchronization program. It synchronizes files between two or more computers in real time, safely protected from prying eyes. Your data is your data alone and you deserve to choose where it is stored, whether it is shared with some third party, and how it's transmitted over the internet.
        tagline:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Free, secure, and distributed file synchronisation tool.
        icon:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Syncthing/icon.png
        screenshot_link:
          type: array
          items:
            type: string
          example:
            - https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/OpenSpeedTest/screenshot-1.png
            - https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/OpenSpeedTest/screenshot-2.png
          x-oapi-codegen-extra-tags:
            mapstructure: "screenshot_link"
            yaml: "screenshot_link,omitempty"
        thumbnail:
          type: string
          example: https://cdn.jsdelivr.net/gh/IceWhaleTech/CasaOS-AppStore@main/Apps/Jellyfin/thumbnail.jpg
        author:
          type: string
          example: CasaOS Team
        developer:
          type: string
          example: Syncthing
        category:
          type: string
          example: Backup
        tips:
          $ref: "#/components/schemas/TipsStoreInfo"
        scheme:
          $ref: "#/components/schemas/Scheme"
        hostname:
          type: string
          example: ""
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        port_map:
          type: string
          example: "8384"
          x-oapi-codegen-extra-tags:
            mapstructure: "port_map"
            yaml: "port_map,omitempty"
        index:
          type: string
          example: "/"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        apps:
          title: store info of container apps included in the compose app
          description: |
            The key is the name of a container app, and the value is the StoreInfo of this container app.
          readOnly: true
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AppStoreInfo"
          example:
            syncthing: {}
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        main:
          title: name of the main service in `services`
          description: |
            The main service is one of the services that the user will interact with.
          type: string
          example: "syncthing"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        architectures:
          title: architectures supported by the compose app
          description: |
            The architectures supported by the compose app. If the architecture of the host is not in this list, the compose app will not be installed.
          type: array
          items:
            type: string
          example:
            - "amd64"
            - "arm64"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        is_uncontrolled: # uncontrolled == not stable and latest
          title: is the app uncontrolled. 
          description: |
            the mean the image tag is custom. not stable or latest
          type: boolean
          example: false
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
    ComposeApp:
      type: object
      description: |-
        See [Compose Specification](https://compose-spec.io) for the schema structure of `ComposeApp`.

      x-go-type: types.Project
      x-go-type-import:
        name: types
        path: github.com/compose-spec/compose-go/types
    ComposeAppWithStoreInfo:
      description: |-
        Schema for `ComposeAppOK` response.

        > Do not mistake it as `ComposeAppStoreInfo` which is just store information of a compose app.
      properties:
        store_info:
          $ref: "#/components/schemas/ComposeAppStoreInfo"
        compose:
          $ref: "#/components/schemas/ComposeApp"
        status:
          type: string
          readOnly: true
          example: running
        update_available:
          readOnly: true
          type: boolean
          example: true
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        is_uncontrolled:
          readOnly: true
          title: is the app uncontrolled. 
          description: |
            the mean the image tag is custom. not stable or latest
          type: boolean
          example: false
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
...
    AppStoreInfo:
      title: StoreInfo of an app
      description: |-
        > Do not mistake it as *information of AppStore* which does not exist.
      required:
        - image
        - envs
        - ports
        - volumes
        - devices
      properties:
        image:
          type: string
          example: "syncthing:latest"
        envs:
          type: array
          items:
            $ref: "#/components/schemas/EnvStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        ports:
          type: array
          items:
            $ref: "#/components/schemas/PortStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        volumes:
          type: array
          items:
            $ref: "#/components/schemas/VolumeStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceStoreInfo"
          x-oapi-codegen-extra-tags:
            yaml: ",omitempty"
...
    EnvStoreInfo:
      required:
        - container
        - description
      properties:
        container:
          type: string
          example: TZ
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Time Zone
    PortStoreInfo:
      required:
        - container
        - description
      properties:
        container:
          type: string
          example: 8384
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Web UI
...
    VolumeStoreInfo:
      required:
        - container
        - description
      properties:
        container:
          type: string
          example: /config
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: Syncthing config directory
    DeviceStoreInfo:
      properties:
        container:
          type: string
          example: /dev/dri
        description:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: GPU device
    TipsStoreInfo:
      required:
        - before_install
      properties:
        before_install:
          type: object
          additionalProperties:
            type: string
          example:
            en_us: |
              Tips for users before install.
          x-oapi-codegen-extra-tags:
            mapstructure: "before_install"
            yaml: "before_install,omitempty"
        custom:
          type: string
          example: "custom tips by user"
    Scheme:
      type: string
      default: http
      enum:
        - http
        - https
      x-oapi-codegen-extra-tags:
        yaml: ",omitempty"
    StoreAppID:
      title: Store app ID
      description: |
        Store app ID of the compose app (unique across all app stores)

        > The `store_app_id` might be same as app name most of the time. However app name might change, due to cloning for example.
        > If app name has changed, there is no way to associate it with the original app in the app store, unless we use the `store_app_id`.
      readOnly: true
      type: string
      example: "syncthing"
      x-go-name: StoreAppID
      x-oapi-codegen-extra-tags:
        mapstructure: "store_app_id"
        yaml: "store_app_id,omitempty"
...